fix: address ruff lint failures across src and tests  - Replace blind 'except Exception' with specific exception types   (BLE001) in build_epub, components, gradio_app, pipeline,   test_compile, unlimited_ocr - Bind loop variables for closure (B023) in pipeline.replace_tag - Auto-fix I001 (import order), UP035 (typing -> collections.abc),   UP037, RUF022 (__all__ sort), RUF046 (unnecessary int()),   RUF100 (unused noqa) - Add noqa: F401 for PyInstaller-driven imports and bitsandbytes   availability probe  Verified with 'ruff check src tests' (0 errors) and AST parse of all modified files.

+++ src/relictoepub/ingest.py
@@ -16,9 +16,9 @@ from __future__ import annotations
 import logging
 import os
 import tempfile
+from collections.abc import Iterator
 from dataclasses import dataclass, field
 from pathlib import Path
-from typing import Iterator
 
 import pymupdf as fitz  # PyMuPDF
 from PIL import Image
@@ -79,8 +79,8 @@ def _normalize_to_square(pil_image: Image.Image, target_size: int) -> Image.Imag
 
     # Scala l'immagine per riempire il lato lungo mantenendo aspect ratio
     scale = target_size / max(w, h)
-    new_w = max(1, int(round(w * scale)))
-    new_h = max(1, int(round(h * scale)))
+    new_w = max(1, round(w * scale))
+    new_h = max(1, round(h * scale))
     resized = pil_image.resize((new_w, new_h), Image.Resampling.LANCZOS)
 
     # Canvas quadrato bianco
